+++ python/GafferUITest/StringPlugValueWidgetTest.py
@@ -91,5 +91,43 @@ class StringPlugValueWidgetTest( GafferUITest.TestCase ) :
 		self.assertEqual( w.textWidget().getText(), "" )
 		self.assertEqual( w.textWidget()._qtWidget().placeholderText(), "test" )
 
+	def testMixedValuesPreserved( self ) :
+
+		n = Gaffer.Node()
+		n["user"]["p1"] = Gaffer.StringPlug()
+		n["user"]["p2"] = Gaffer.StringPlug()
+
+		n["user"]["p1"].setValue( "p1" )
+		n["user"]["p2"].setValue( "p2" )
+
+		Gaffer.Metadata.registerValue( n["user"]["p1"], "stringPlugValueWidget:placeholderText", "test" )
+		Gaffer.Metadata.registerValue( n["user"]["p2"], "stringPlugValueWidget:placeholderText", "test" )
+
+		w = GafferUI.StringPlugValueWidget( { n["user"]["p1"], n["user"]["p2"] } )
+		self.assertEqual( w.textWidget().getText(), "" )
+		self.assertEqual( w.textWidget()._qtWidget().placeholderText(), "---" )
+
+		w.textWidget()._qtWidget().editingFinished.emit()
+
+		self.assertEqual( n["user"]["p1"].getValue(), "p1" )
+		self.assertEqual( n["user"]["p2"].getValue(), "p2" )
+
+		self.assertEqual( w.textWidget().getText(), "" )
+		self.assertEqual( w.textWidget()._qtWidget().placeholderText(), "---" )
+
+		# Simulate the user editing, even if it results in an empty string
+		self.assertEqual( w.textWidget()._qtWidget().text(), "" )
+		w.textWidget()._qtWidget().textChanged.emit( w.textWidget()._qtWidget().text() )
+
+		self.assertEqual( w.textWidget()._qtWidget().placeholderText(), "test" )
+
+		self.assertEqual( n["user"]["p1"].getValue(), "p1" )
+		self.assertEqual( n["user"]["p2"].getValue(), "p2" )
+
+		w.textWidget()._qtWidget().editingFinished.emit()
+
+		self.assertEqual( n["user"]["p1"].getValue(), "" )
+		self.assertEqual( n["user"]["p2"].getValue(), "" )
+
 if __name__ == "__main__":
 	unittest.main()