+++ Changes.md
@@ -6,7 +6,7 @@ Fixes
 
 - Viewer : Fixed bug that caused the Inspector to grow very wide when inspecting long strings.
 - NumericPlugValueWidget : Fixed bug that caused plug values to be set to 0 if editing completed with invalid text.
-- StringPlugValueWidget : Fixed bug that could cause plug values to be overwritten when editing multiple plugs and the value state changed to mixed via external action.
+- StringPlugValueWidget : Fixed bug that could cause plug values to be overwritten when editing multiple plugs and focus was lost without change, or the value state changed to mixed via external action.
 
 API
 ---